Simplify types definitions.

+++ docs/sql/mysql.sql
@@ -33,7 +33,7 @@ CREATE TABLE `daemons` (
 -- 
 
 CREATE TABLE `gammu` (
-  `Version` tinyint(4) NOT NULL default '0'
+  `Version` integer NOT NULL default '0'
 ) ENGINE=MyISAM DEFAULT CHARSET=utf8;
 
 -- 
@@ -56,9 +56,9 @@ CREATE TABLE `inbox` (
   `Coding` enum('Default_No_Compression','Unicode_No_Compression','8bit','Default_Compression','Unicode_Compression') NOT NULL default 'Default_No_Compression',
   `UDH` text NOT NULL,
   `SMSCNumber` varchar(20) NOT NULL default '',
-  `Class` int(11) NOT NULL default '-1',
+  `Class` integer NOT NULL default '-1',
   `TextDecoded` varchar(160) NOT NULL default '',
-  `ID` int(11) unsigned NOT NULL auto_increment,
+  `ID` integer unsigned NOT NULL auto_increment,
   `RecipientID` text NOT NULL,
   `Processed` enum('false','true') NOT NULL default 'false',
   UNIQUE KEY `ID` (`ID`)
@@ -83,11 +83,11 @@ CREATE TABLE `outbox` (
   `DestinationNumber` varchar(20) NOT NULL default '',
   `Coding` enum('Default_No_Compression','Unicode_No_Compression','8bit','Default_Compression','Unicode_Compression') default '8bit',
   `UDH` text,
-  `Class` int(11) default '-1',
+  `Class` integer default '-1',
   `TextDecoded` varchar(160) NOT NULL default '',
-  `ID` int(11) unsigned NOT NULL auto_increment,
+  `ID` integer unsigned NOT NULL auto_increment,
   `MultiPart` enum('false','true') default 'false',
-  `RelativeValidity` int(11) default '-1',
+  `RelativeValidity` integer default '-1',
   `SenderID` text,
   `SendingTimeOut` timestamp NULL default '0000-00-00 00:00:00',
   `DeliveryReport` enum('default','yes','no') default 'default',
@@ -110,10 +110,10 @@ CREATE TABLE `outbox_multipart` (
   `Text` text,
   `Coding` enum('Default_No_Compression','Unicode_No_Compression','8bit','Default_Compression','Unicode_Compression') default '8bit',
   `UDH` text,
-  `Class` int(11) default '-1',
+  `Class` integer default '-1',
   `TextDecoded` varchar(160) default NULL,
-  `ID` int(11) unsigned NOT NULL default '0',
-  `SequencePosition` int(11) NOT NULL default '1'
+  `ID` integer unsigned NOT NULL default '0',
+  `SequencePosition` integer NOT NULL default '1'
 ) ENGINE=MyISAM DEFAULT CHARSET=utf8;
 
 -- 
@@ -128,7 +128,7 @@ CREATE TABLE `outbox_multipart` (
 -- 
 
 CREATE TABLE `pbk` (
-  `GroupID` int(11) NOT NULL default '-1',
+  `GroupID` integer NOT NULL default '-1',
   `Name` text NOT NULL,
   `Number` text NOT NULL
 ) ENGINE=MyISAM DEFAULT CHARSET=utf8;
@@ -146,7 +146,7 @@ CREATE TABLE `pbk` (
 
 CREATE TABLE `pbk_groups` (
   `Name` text NOT NULL,
-  `ID` int(11) NOT NULL auto_increment,
+  `ID` integer NOT NULL auto_increment,
   UNIQUE KEY `ID` (`ID`)
 ) ENGINE=MyISAM DEFAULT CHARSET=utf8 AUTO_INCREMENT=1 ;
 
@@ -170,8 +170,8 @@ CREATE TABLE `phones` (
   `Receive` enum('yes','no') NOT NULL default 'no',
   `IMEI` text NOT NULL,
   `Client` text NOT NULL,
-  `Battery` int(2) NOT NULL,
-  `Signal` int(2) NOT NULL,
+  `Battery` integer NOT NULL,
+  `Signal` integer NOT NULL,
   `Sent` int NOT NULL,
   `Received` int NOT NULL
 ) ENGINE=MyISAM DEFAULT CHARSET=utf8;
@@ -196,15 +196,15 @@ CREATE TABLE `sentitems` (
   `Coding` enum('Default_No_Compression','Unicode_No_Compression','8bit','Default_Compression','Unicode_Compression') NOT NULL default '8bit',
   `UDH` text NOT NULL,
   `SMSCNumber` varchar(20) NOT NULL default '',
-  `Class` int(11) NOT NULL default '-1',
+  `Class` integer NOT NULL default '-1',
   `TextDecoded` varchar(160) NOT NULL default '',
-  `ID` int(11) unsigned NOT NULL default '0',
+  `ID` integer unsigned NOT NULL default '0',
   `SenderID` text NOT NULL,
-  `SequencePosition` int(11) NOT NULL default '1',
+  `SequencePosition` integer NOT NULL default '1',
   `Status` enum('SendingOK','SendingOKNoReport','SendingError','DeliveryOK','DeliveryFailed','DeliveryPending','DeliveryUnknown','Error') NOT NULL default 'SendingOK',
-  `StatusError` int(11) NOT NULL default '-1',
-  `TPMR` int(11) NOT NULL default '-1',
-  `RelativeValidity` int(11) NOT NULL default '-1',
+  `StatusError` integer NOT NULL default '-1',
+  `TPMR` integer NOT NULL default '-1',
+  `RelativeValidity` integer NOT NULL default '-1',
   `CreatorID` text NOT NULL
 ) ENGINE=MyISAM DEFAULT CHARSET=utf8;
